ai_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
google-generativeai
langchain
langchain-google-genai
